@@ -3,9 +3,8 @@ import 'package:flutter_transaction_tracker/domain/transaction.dart';
 
 class TransactionItem extends StatelessWidget {
   final Transaction transaction;
-  const TransactionItem({Key key, @required this.transaction})
-      : assert(transaction != null),
-        super(key: key);
+  const TransactionItem({Key? key, required this.transaction})
+      : super(key: key);
   @override
   Widget build(BuildContext context) {
     return ListTile(
@@ -20,7 +19,8 @@ class TransactionItem extends StatelessWidget {
 class TransactionIndicator extends StatelessWidget {
   final TransactionType transactionType;
 
-  const TransactionIndicator({Key key, this.transactionType}) : super(key: key);
+  const TransactionIndicator({Key? key, required this.transactionType})
+      : super(key: key);
   @override
   Widget build(BuildContext context) {
     final indicatorColor =
