@@ -3,16 +3,18 @@ import 'package:flutter_transaction_tracker/domain/transaction.dart';
 import 'package:flutter_transaction_tracker/utils.dart';
 
 typedef TransactionVoidCallback = void Function(Transaction transaction);
-typedef TransactionTypeCallback = void Function(TransactionType type);
+typedef TransactionTypeCallback = void Function(TransactionType? type);
+typedef ValidatorType = String? Function(String? value);
 
 class TransactionModal extends StatelessWidget {
   final TransactionVoidCallback onButtonPressed;
 
-  const TransactionModal({Key key, this.onButtonPressed}) : super(key: key);
+  const TransactionModal({Key? key, required this.onButtonPressed})
+      : super(key: key);
 
   static Future<void> show(
     BuildContext context, {
-    TransactionVoidCallback onButtonPressed,
+    required TransactionVoidCallback onButtonPressed,
     bool barrierDismissible = true,
   }) {
     return showDialog(
@@ -25,7 +27,7 @@ class TransactionModal extends StatelessWidget {
           padding: const EdgeInsets.all(8.0),
           child: TransactionModal(
             onButtonPressed: (transaction) {
-              onButtonPressed?.call(transaction);
+              onButtonPressed.call(transaction);
               Navigator.pop(context);
             },
           ),
@@ -71,7 +73,7 @@ class TransactionModal extends StatelessWidget {
                 label: 'Description',
                 controller: descriptionController,
                 validator: (value) =>
-                    value.isNotEmpty ? null : 'This field cant be empty',
+                    value!.isNotEmpty ? null : 'This field cant be empty',
               ),
               const SizedBox(height: 8),
               TransactionTextFormField(
@@ -79,7 +81,7 @@ class TransactionModal extends StatelessWidget {
                 inputType: TextInputType.number,
                 controller: amountController,
                 validator: (value) {
-                  if (value.isNotEmpty) {
+                  if (value!.isNotEmpty) {
                     return isNumeric(value) ? null : 'The value is not numeric';
                   } else {
                     return 'This field cant be empty';
@@ -92,7 +94,7 @@ class TransactionModal extends StatelessWidget {
                 inputType: TextInputType.datetime,
                 controller: dateController,
                 validator: (value) =>
-                    value.isNotEmpty ? null : 'This field cant be empty',
+                    value!.isNotEmpty ? null : 'This field cant be empty',
               ),
               const SizedBox(height: 8),
               const SizedBox(height: 16),
@@ -101,7 +103,7 @@ class TransactionModal extends StatelessWidget {
                   icon: const Icon(Icons.save),
                   label: const Text('Save'),
                   onPressed: () {
-                    if (_formKey.currentState.validate()) {
+                    if (_formKey.currentState!.validate()) {
                       final amountText = amountController.text.trim();
                       final dateText = dateController.text.trim();
                       final descriptionText = descriptionController.text.trim();
@@ -126,16 +128,18 @@ class TransactionModal extends StatelessWidget {
 
 class TransactionTextFormField extends StatelessWidget {
   const TransactionTextFormField({
-    Key key,
-    @required this.label,
-    this.controller,
+    Key? key,
+    required this.label,
+    required this.controller,
     this.inputType = TextInputType.text,
     this.validator,
   }) : super(key: key);
+
   final String label;
   final TextEditingController controller;
   final TextInputType inputType;
-  final Function(String) validator;
+  final ValidatorType? validator;
+
   @override
   Widget build(BuildContext context) {
     return TextFormField(
@@ -155,7 +159,8 @@ class TransactionTextFormField extends StatelessWidget {
 }
 
 class TransactionTypeRadioForm extends StatefulWidget {
-  const TransactionTypeRadioForm({Key key, this.callback}) : super(key: key);
+  const TransactionTypeRadioForm({Key? key, required this.callback})
+      : super(key: key);
   final TransactionTypeCallback callback;
 
   @override
@@ -164,7 +169,7 @@ class TransactionTypeRadioForm extends StatefulWidget {
 }
 
 class _TransactionTypeRadioFormState extends State<TransactionTypeRadioForm> {
-  TransactionType _type = TransactionType.income;
+  TransactionType? _type = TransactionType.income;
 
   @override
   Widget build(BuildContext context) {
